chore(storage): remove useless rocksdb mock
Otherwise this would be really painful to port to the latest version.
