+++ node/src/storage_mngr.rs
@@ -366,11 +366,10 @@ pub fn create_appropriate_backend(
         }
         config::StorageBackend::RocksDB => {
             let path = conf.db_path.as_path();
-            let mut options = backends::rocksdb::Options::default();
+            let mut options = backends::rocks::Options::default();
             options.create_if_missing(true);
             options.set_max_open_files(conf.max_open_files);
-            let db =
-                backends::rocksdb::Backend::open(&options, path).map_err(|e| as_failure!(e))?;
+            let db = backends::rocks::Backend::open(&options, path).map_err(|e| as_failure!(e))?;
 
             wrap_storage_as_nodestorage(db, conf)
         }
